+++ apps/mobile/src/components/common/icon.tsx
@@ -9,6 +9,7 @@ import ArrowFatDownBold from 'react-native-phosphor/src/bold/ArrowFatDown'
 import ArrowFatUpBold from 'react-native-phosphor/src/bold/ArrowFatUp'
 import ArrowLeftBold from 'react-native-phosphor/src/bold/ArrowLeft'
 import ArrowRightBold from 'react-native-phosphor/src/bold/ArrowRight'
+import ArrowsInLineVerticalBold from 'react-native-phosphor/src/bold/ArrowsInLineVertical'
 import ArrowsOutBold from 'react-native-phosphor/src/bold/ArrowsOut'
 import ArrowsSplitBold from 'react-native-phosphor/src/bold/ArrowsSplit'
 import ArrowUpBold from 'react-native-phosphor/src/bold/ArrowUp'
@@ -106,6 +107,7 @@ import ArrowFatDownDuotone from 'react-native-phosphor/src/duotone/ArrowFatDown'
 import ArrowFatUpDuotone from 'react-native-phosphor/src/duotone/ArrowFatUp'
 import ArrowLeftDuotone from 'react-native-phosphor/src/duotone/ArrowLeft'
 import ArrowRightDuotone from 'react-native-phosphor/src/duotone/ArrowRight'
+import ArrowsInLineVerticalDuotone from 'react-native-phosphor/src/duotone/ArrowsInLineVertical'
 import ArrowsOutDuotone from 'react-native-phosphor/src/duotone/ArrowsOut'
 import ArrowsSplitDuotone from 'react-native-phosphor/src/duotone/ArrowsSplit'
 import ArrowUpDuotone from 'react-native-phosphor/src/duotone/ArrowUp'
@@ -203,6 +205,7 @@ import ArrowFatDownFill from 'react-native-phosphor/src/fill/ArrowFatDown'
 import ArrowFatUpFill from 'react-native-phosphor/src/fill/ArrowFatUp'
 import ArrowLeftFill from 'react-native-phosphor/src/fill/ArrowLeft'
 import ArrowRightFill from 'react-native-phosphor/src/fill/ArrowRight'
+import ArrowsInLineVerticalFill from 'react-native-phosphor/src/fill/ArrowsInLineVertical'
 import ArrowsOutFill from 'react-native-phosphor/src/fill/ArrowsOut'
 import ArrowsSplitFill from 'react-native-phosphor/src/fill/ArrowsSplit'
 import ArrowUpFill from 'react-native-phosphor/src/fill/ArrowUp'
@@ -300,6 +303,7 @@ import ArrowFatDownRegular from 'react-native-phosphor/src/regular/ArrowFatDown'
 import ArrowFatUpRegular from 'react-native-phosphor/src/regular/ArrowFatUp'
 import ArrowLeftRegular from 'react-native-phosphor/src/regular/ArrowLeft'
 import ArrowRightRegular from 'react-native-phosphor/src/regular/ArrowRight'
+import ArrowsInLineVerticalRegular from 'react-native-phosphor/src/regular/ArrowsInLineVertical'
 import ArrowsOutRegular from 'react-native-phosphor/src/regular/ArrowsOut'
 import ArrowsSplitRegular from 'react-native-phosphor/src/regular/ArrowsSplit'
 import ArrowUpRegular from 'react-native-phosphor/src/regular/ArrowUp'
@@ -501,6 +505,12 @@ const icons = {
     duotone: ArrowUpDuotone,
     fill: ArrowUpFill,
     regular: ArrowUpRegular,
+  },
+  ArrowsInLineVertical: {
+    bold: ArrowsInLineVerticalBold,
+    duotone: ArrowsInLineVerticalDuotone,
+    fill: ArrowsInLineVerticalFill,
+    regular: ArrowsInLineVerticalRegular,
   },
   ArrowsOut: {
     bold: ArrowsOutBold,